general: little clean-up in cleanup scripts
  * we do not need to set THIS_TEST_NAME, it is done in settings
  * unify the reporting, use the "overall" format (with '#')
  * skip clean-up when the tests are run elsewhere
    - we do not want to do clean-up in a dir specified by a variable,
      since it is unsafe -- what if the variable contains '/'...
    - there is no need for deleting logs in thetestsuite tree when they
      are not present there; it is more fair to report "SKIP" in such
      cases

+++ base_annotate/cleanup.sh
@@ -10,13 +10,15 @@
 . ../common/init.sh
 . ./settings.sh
 
-THIS_TEST_NAME=`basename $0 .sh`
+if [ -n "$PERFSUITE_RUN_DIR" ]; then
+	print_overall_skipped
+	exit 0
+fi
 
 make -s -C examples clean
-
 find . -name \*.log | xargs -r rm
 find . -name \*.err | xargs -r rm
 rm -f perf.data*
 
-print_results 0 0 "clean-up logs"
+print_overall_results 0
 exit $?
