+++ base_trace/cleanup.sh
@@ -12,6 +12,11 @@
 . ../common/init.sh
 . ./settings.sh
 
+if [ -n "$PERFSUITE_RUN_DIR" ]; then
+	print_overall_skipped
+	exit 0
+fi
+
 find . -name \*.log | xargs -r rm
 find . -name \*.err | xargs -r rm
 rm -f perf.data*
